requirements.txt
# Flask and Web Framework
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Machine Learning / Technical Analysis
scikit-learn==1.3.0
tensorflow==2.13.0
ta==0.11.0

# NLP / Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.0
apscheduler